examples/bug_bounty/ultra_gemini_production/requirements.txt
google-generativeai>=0.3.0
psutil>=5.9.0
asyncio-throttle>=1.0.0
aiohttp>=3.8.0
aiofiles>=0.8.0
pyyaml>=6.0
fastapi>=0.100.0
uvicorn>=0.20.0
prometheus-client>=0.15.0
